TBTables for org and user types

+++ sql/procedures/Config.sql
@@ -98,6 +98,7 @@ call ValidateOrgSuperUser(i_ses, 0, v_me);
 if i_newid is not null then
 	update tasks set status=i_newid where status=i_id;
 end if;
+
 delete from taskstatus where statusid=i_id;
 
 if row_count() < 1 then
@@ -168,6 +169,7 @@ create procedure DeleteUserStatus (
 ) begin
 
 declare v_me int unsigned;
+declare v_cnt int unsigned;
 
 declare exit handler for sqlexception begin
 	rollback;
@@ -176,10 +178,98 @@ end;
 
 call ValidateUserMaintRights(i_ses, 0, v_me);
 
-delete from userstatus where statusid=i_id;
+select count(*) into v_cnt from users where status=i_id;
+if v_cnt > 0 then
+	select -1452 into @err;
+else
+
+	delete from userstatus where statusid=i_id;
+
+	if row_count() < 1 then
+		select -1451 into @err;
+	end if;
+end if;
+
+end$$
+
+DROP PROCEDURE IF EXISTS CreateOrgType$$
+
+create procedure CreateOrgType (
+	i_ses	bigint unsigned,
+	i_name	varchar(40)
+) begin
+
+declare v_me int unsigned;
+
+declare exit handler for sqlexception begin
+	rollback;
+	select -1460 into @err;
+end;
+
+call ValidateUserMaintRights(i_ses, 0, v_me);
+
+insert into orgstatus (name) values (i_name);
 
 if row_count() < 1 then
-	select -1451 into @err;
+	select -1461 into @err;
+end if;
+
+end$$
+
+
+DROP PROCEDURE IF EXISTS UpdateOrgType$$
+
+create procedure UpdateOrgType (
+	i_ses	bigint unsigned,
+	i_id	int unsigned,
+	i_name	varchar(40)
+) begin
+
+declare v_me int unsigned;
+
+declare exit handler for sqlexception begin
+	rollback;
+	select -1470 into @err;
+end;
+
+call ValidateUserMaintRights(i_ses, 0, v_me);
+
+update orgstatus set name=i_name where statusid=i_id;
+
+if row_count() < 1 then
+	select -1471 into @err;
+end if;
+
+end$$
+
+
+DROP PROCEDURE IF EXISTS DeleteOrgType$$
+
+create procedure DeleteOrgType (
+	i_ses	bigint unsigned,
+	i_id	int unsigned
+) begin
+
+declare v_me int unsigned;
+declare v_cnt int unsigned;
+
+declare exit handler for sqlexception begin
+	rollback;
+	select -1480 into @err;
+end;
+
+call ValidateUserMaintRights(i_ses, 0, v_me);
+
+select count(*) into v_cnt from organizations where status=i_id;
+if v_cnt > 0 then
+	select -1482 into @err;
+else
+
+	delete from orgstatus where statusid=i_id;
+
+	if row_count() < 1 then
+		select -1481 into @err;
+	end if;
 end if;
 
 end$$
